refactor: 예외처리 관 패키지 구조 및 경로 변경
