+++ src/main/java/org/ject/support/admin/auth/service/AdminAuthService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.ject.support.common.security.jwt.JwtTokenProvider;
 import org.ject.support.common.util.CodeGeneratorUtil;
 import org.ject.support.admin.auth.component.AdminMemberComponent;
-import org.ject.support.domain.admin.exception.AdminErrorCode;
-import org.ject.support.domain.admin.exception.AdminException;
+import org.ject.support.admin.exception.AdminErrorCode;
+import org.ject.support.admin.exception.AdminException;
 import org.ject.support.domain.member.MemberStatus;
 import org.ject.support.domain.member.entity.Member;
 import org.ject.support.external.infrastructure.SlackRateLimiter;
