+++ src/main/java/org/ject/support/admin/exception/AdminException.java
@@ -1,4 +1,4 @@
-package org.ject.support.domain.admin.exception;
+package org.ject.support.admin.exception;
 
 import org.ject.support.common.exception.BusinessException;
